Made some refactoring over admin manager

+++ src/main/java/com/ita/edu/softserve/manager/AdminManager.java
@@ -14,6 +14,6 @@ public interface AdminManager extends BaseManager {
 	public void printAllUsers();
 
 	/** Prints the number of all users */	 
-	public Integer countAllUsers();
+	public Long countAllUsers();
 
 }
